requirements.txt
fastapi[standard]
pyjwt
jinja2
google-cloud-firestore==2.20.1
pydantic-settings
python-multipart
google-cloud-storage